Refactoring parametru
Nacitani parametru presunuto do funkceSpolecne.cpp stejne tak navratove
hodnoty.
Zlepsene nacitani integeru pomoci istringstream misto atoi().

Timto prohlasuji vetev 04 za uzavrenou
